fix(live): try to fix questions in loop
When teacher go forvard before student and come to node which needs waitin on students side, student's JS will post data to server every 2 sec.
In case if we are in loop we don't need to actually save this data, just return message to the user (msg could be new one or old one).