fix(join_course_redirect): Fixed join course loogic.
After clicking on link from email user will be redirected to join course page which will redirect user to shared courses if user is logged in. If user is not logged in it will add u_hash to users session and show SignupForm with pre-fill email in no such user registered with email from invite, if there's user with the same email - it will show LoginForm. Next page will be shared_courses.